Refactor subscription procedure to match e2 submgr; make KPM subscriptions.
There's a lot of conflation of RMR and e2 subscriptions.  Basically, you
must send a sub req with subid=-1 so that RIC knows how to process it.
(In my accidental experimentation, subid is ignored, but who knows.)
Then, if you want to be able to match your sub req up with a sub resp,
you *must* set the RMR transaction id (xid, aka xact) header field;
the sub resp then carries this xid.  Otherwise, there is no way to match;
the RIC changes the requestor_id and instance_id bits in your original
sub req before passing it to the RAN node, so that it can aggregate.
Anyway, the sub resp carries the "real" subid in the RMR header, and
that is the thing you want to match indications against.

So now we have a two-part process where we keep a map of transaction id
-> pending_subscriptions, and successful subscription responses then
match in this map and are removed; and then placed into the regular
subscriptions map using the new subid as the key.

This required that we fork the xapp-frame-cpp codebase to add
accessors for the xid (aka xact, if you're RMR) header field.